requirements_fred.txt
fredapi>=0.5.2
pandas>=2.0
pyarrow>=14.0
pyyaml>=6.0
